+++ lib/speed_up.dart
@@ -1,6 +1,7 @@
 library speed_up;
 
 export 'src/collection_extensions.dart';
+export 'src/enum_extensions.dart';
 export 'src/funs.dart';
 export 'src/int_extensions.dart';
 export 'src/num_extensions.dart';