+++ lib/src/enum_extensions.dart
@@ -1,0 +1,3 @@
+extension EnumExtensions on Enum? {
+  String get title => toString().split('.').last;
+}