+++ test/int_extensions_test.dart
@@ -27,4 +27,34 @@ void main() {
       });
     });
   });
+
+  given('epoch', () {
+    const epoch = 1636865303;
+
+    then('toDateFromSecUnixTimestamp in UTC should retrun Date', () {
+      final date = epoch.toDateFromSecUnixTimestamp();
+
+      date.year.should.be(2021);
+      date.month.should.be(11);
+      date.day.should.be(14);
+      date.hour.should.be(4);
+      date.minute.should.be(48);
+      date.second.should.be(23);
+    });
+
+    then('toDateFromSecUnixTimestamp in Local should retrun Date', () {
+      final date = epoch.toDateFromSecUnixTimestamp(asLocal: true);
+
+      final utc = epoch.toDateFromSecUnixTimestamp();
+      final offset = DateTime.now().timeZoneOffset;
+      final local = utc.add(offset);
+
+      date.year.should.be(local.year);
+      date.month.should.be(local.month);
+      date.day.should.be(local.day);
+      date.hour.should.be(local.hour);
+      date.minute.should.be(local.minute);
+      date.second.should.be(local.second);
+    });
+  });
 }